Consolidated TagServices, updated allPosts link, and fixed some merge issues
